Refactor: For Test
