+++ src/main/java/capstone/relation/websocket/signaling/dto/SdpResponseDto.java
@@ -2,9 +2,12 @@ package capstone.relation.websocket.signaling.dto;
 
 import capstone.relation.user.dto.UserInfoDto;
 import io.swagger.v3.oas.annotations.media.Schema;
+import lombok.Builder;
 import lombok.Data;
+import lombok.NoArgsConstructor;
 
 @Data
+@NoArgsConstructor
 @Schema(description = "offer 또는 answer를 받을 때 사용하는 DTO")
 public class SdpResponseDto {
 	@Schema(description = "메시지를 보낸 사용자 정보(즉 자신이 아닌 통신하고 있는 유저 id)")
@@ -15,4 +18,11 @@ public class SdpResponseDto {
 	@Schema(description = "메시지 타입 \n"
 		+ "ScreenShare | Video", example = "ScreenShare")
 	private SignalMessageType type;
+
+	@Builder
+	public SdpResponseDto(UserInfoDto userInfo, SdpDto sessionDescription, SignalMessageType type) {
+		this.userInfo = userInfo;
+		this.sessionDescription = sessionDescription;
+		this.type = type;
+	}
 }
